quest: Allow ItemId enum for quest files

+++ Arrowgene.Ddon.GameServer/Handler/QuestGetRewardBoxItemHandler.cs
@@ -64,7 +64,7 @@ namespace Arrowgene.Ddon.GameServer.Handler
 
             var distinctRewards = packet.GetRewardBoxItemList.Select(x => x.UID).Distinct().ToList();
 
-            var slotCount = coalescedRewards.Sum(x => distinctRewards.Contains(x.Key) ? Server.ItemManager.PredictAddItemSlots(client.Character, StorageType.StorageBoxNormal, x.Value.ItemId, x.Value.Num) : 0);
+            var slotCount = coalescedRewards.Sum(x => distinctRewards.Contains(x.Key) ? Server.ItemManager.PredictAddItemSlots(client.Character, StorageType.StorageBoxNormal, (uint) x.Value.ItemId, x.Value.Num) : 0);
             if (slotCount > client.Character.Storage.GetStorage(StorageType.StorageBoxNormal).EmptySlots())
             {
                 throw new ResponseErrorException(ErrorCode.ERROR_CODE_ITEM_STORAGE_OVERFLOW);
@@ -83,7 +83,7 @@ namespace Arrowgene.Ddon.GameServer.Handler
                     }
                     else if (reward.Num > 0)
                     {
-                        var result = Server.ItemManager.AddItem(Server, client.Character, false, reward.ItemId, reward.Num, connectionIn: connection);
+                        var result = Server.ItemManager.AddItem(Server, client.Character, false, (uint) reward.ItemId, reward.Num, connectionIn: connection);
                         updateCharacterItemNtc.UpdateItemList.AddRange(result);
                     }
                 }
